requirements.txt
streamlit
langchain
langchain-community
langchain-groq
python-dotenv
sentence_transformers
beautifulsoup4
faiss-cpu
